requirements.txt
transformers
accelerate
bitsandbytes
peft
datasets
trl
pyyaml
